Fix bug with distance calculation
In contrast to what it was before we have to calculate distance
keeping in mind message direction.

Example. We have nodes 0, 1, 2. Node 1 is dead. Node 0 sends message
to the right first to 1, then to 2. Old distance between 0 and 2 is 1,
but true distance is two, because we should take into account that
messages are directed.