Change phase end calculation to recursive algorithm
This algorithm is supposed to give a correct answer even for non-full
trees.

Nevertheless I found a problem. What if g > o, what should determine
the end of the phase o or g? If o, then the last node, can't send the
correction message timely. If g, then we the rest just wastes time not
sending the message.