Implement optimal tree broadcast
This version does not implement interleaving

I actually implemented two version of optimal tree. The first version
is basically commented out. It is correct, but it is much slower. It
produces slighly different result, because for the case of not
complete tree, it makes nodes with higher to send additional messages.

The second implemntation is less obvious, but much faster. And also it
renders slighly different, although equivalent result.